[IMP] delivery_procurement_group_carrier: Refactor Unittests add Common class
